Let redux throw if the action is malformed
This lib is written in typescript and according to the type definitions
it should spot if you are returning something else than an action inside
your stream processing. The error could be catched and returned to the
stream for it to handle but because it is often handled as:
```
stream$.replaceError(() => xs.empty());
```
Or something else in JS without regard to the error it could lead to
unforeseen consequence as well as infinite loops if the strean is
attempted to be restarted.
Instead I align to the `redux-observable` behaviour by not trying to
catch this error and letting it bubble outside the stream.