Update dependencies
Update *all&* dependencies to latest versions.
Babel is not updated to babel 7 yet.
A small update was needed in the type signatures of `createMiddleware`
because of changes in the redux api (it now take into accounts the
possible overloading of dispatch when using middlewares, those changes
are reflected in the api although it is not supported inside this
package. Users should be careful as to what is dispatched in redux
anyway.